src/test/javascript/spec/app/entities/voiture/voiture.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { VoitureService } from 'app/entities/voiture/voiture.service';
import { IVoiture, Voiture } from 'app/shared/model/voiture.model';
import { Typevehicule } from 'app/shared/model/enumerations/typevehicule.model';

describe('Service Tests', () => {
  describe('Voiture Service', () => {
    let injector: TestBed;
    let service: VoitureService;
    let httpMock: HttpTestingController;
    let elemDefault: IVoiture;
    let expectedResult: IVoiture | IVoiture[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule]
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(VoitureService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new Voiture(0, 'AAAAAAA', Typevehicule.VOITURE, 0);
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a Voiture', () => {
        const returnedFromService = Object.assign(
          {
            id: 0
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new Voiture()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a Voiture', () => {
        const returnedFromService = Object.assign(
          {
            matricule: 'BBBBBB',
            type: 'BBBBBB',
            capacite: 1
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of Voiture', () => {
        const returnedFromService = Object.assign(
          {
            matricule: 'BBBBBB',
            type: 'BBBBBB',
            capacite: 1
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a Voiture', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
